move `new-external-account` to `OtherAccounts` provider
